Combined java classes for brevity
