+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/autonomous/OneStone.java
@@ -1,4 +1,6 @@
-package org.firstinspires.ftc.teamcode.opmode.autonomous.onestone;
+package org.firstinspires.ftc.teamcode.opmode.autonomous;
+
+import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 
 import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 import org.firstinspires.ftc.teamcode.opmode.AutonomousOpMode;
@@ -53,5 +55,38 @@ public class OneStone extends AutonomousOpMode {
 
         driveY(10, 0.5);
 
+    }
+
+    @Autonomous(name="Blue One Stone - Bridge", group="Blue")
+    public class BlueBridge extends OneStone {
+
+        public void runOpMode(){
+            runOneStone(TeamColor.BLUE, ParkSide.BRIDGE);
+        }
+    }
+
+    @Autonomous(name="Blue One Stone - Wall", group="Blue")
+    public class BlueWall extends OneStone {
+        public void runOpMode(){
+            runOneStone(TeamColor.BLUE, ParkSide.WALL);
+        }
+    }
+
+    @Autonomous(name="Red One Stone - Bridge", group="Red")
+    public class RedBridge extends OneStone {
+
+        public void runOpMode(){
+            runOneStone(TeamColor.RED, ParkSide.BRIDGE);
+
+        }
+    }
+
+    @Autonomous(name="Red One Stone - Wall", group="Red")
+    public class RedWall extends OneStone {
+    
+        public void runOpMode(){
+            runOneStone(TeamColor.RED, ParkSide.WALL);
+
+        }
     }
 }
